Release 0.2.0
Squashed commit of the following:

commit d68407f104d1d909cd102779f604626ecb95aaa6
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 27 13:58:31 2020 +0100

    Bump version number

commit a69bb7e25a0db0aa9104c3b3011e223762ed567c
Merge: 10b9f2a 6b3c1df
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 27 13:51:03 2020 +0100

    Merge branch 'master' into develop

commit 10b9f2a4df1e71794281a72d037745976cabeb7c
Merge: 72d9fd5 3f5ba8a
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 27 13:45:25 2020 +0100

    Merge branch 'feature/nonconvex_phase' in develop

commit 3f5ba8a566afc2feaeee3e436a5fcd2701ab127d
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 27 13:42:46 2020 +0100

    feat(nonconvex): nonconvex EVM implementation and logging tools

commit d34429e5b7b1e93f430912e2a7080702b59517db
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 20 21:35:37 2020 +0100

    #1 tavis file naming was wrong?

commit 1603de0e5ba9bad2e5a40396137b5698e7a78844
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 20 21:33:17 2020 +0100

    #1 trigger build hopefully

commit cc06df64adbb2df67e7d00dbc583b1d99430188c
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 20 21:29:09 2020 +0100

    refactor(CI): update travis config

    Adapted travis.yml file from pylhc ones as template, hopefully all is good.

commit 05e928dafd3565fc8c659bfa275b93c84656e712
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 20 21:20:08 2020 +0100

    refactor(content): logging, contexts and pylint

    Quite a consequent overhaul including the following:
     - Incorporated `logging_tools` and `contexts` utility modules from omc3 in order to not have to import it, since it is very consequent;
     - Refactored files to make use of those tools;
     - Added comments to tell pylint to fuck off in places I know things are ok.

commit 6df68aed9266d96e8eb01b39458b197bd93ec2e9
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Feb 19 16:23:00 2020 +0100

    you know it

commit 812ca2a246c02789e6afbcd659c5b7dc80f44a15
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Mon Feb 17 09:19:37 2020 +0100

    yoou know it

commit d6f3822b6d2ea0ece4d671ef0eaccb5139f26ff4
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Feb 12 17:42:01 2020 +0100

    quick update

commit 6f8ac1114dd4f74df27718c78d341bd7660a723c
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Fri Feb 7 14:05:08 2020 +0100

    update

commit c4a89d6a7c5dd87fa9f1fce5d97a30af7a18bb46
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Fri Feb 7 10:21:08 2020 +0100

    quick update

commit 3efbb24ed0108d12e40c879e7bb45fcc9e48a573
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 6 18:38:59 2020 +0100

    quick update

commit d15f5c0f609bf90bd079469b826093903e620130
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 6 17:25:40 2020 +0100

    quick update

commit 5963a50c9a006e824642a9503ff45ae761ffb967
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 6 16:54:51 2020 +0100

    quick update

commit 498b46e52f124c0eced24056197e5bfcdc716af0
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 6 16:48:31 2020 +0100

    dump

commit bc3fdf62079edf41253162ad70fab9d9e91b19bb
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Thu Feb 6 12:11:18 2020 +0100

    dump

commit 1e7e469189f79aa23a6a461111922cfd8699072b
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Feb 5 17:55:34 2020 +0100

    dump

commit 1ab0718157f3c258664d86c9182263b5403f1828
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Mon Feb 3 19:02:11 2020 +0100

    quick update

commit 380c5e0581e9eb272b8aae0044cae50a6f47f68c
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Fri Jan 31 18:01:05 2020 +0100

    include saving into plotting functions

commit 532b505becbd82de08d24475d9548b98856da730
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Jan 29 12:08:08 2020 +0100

    refactor(content): plotting style

    Updated PLOT_PARAMS with much more settings
    Now used in nonconvex
    Makefile doesn't call isort anymore, should be done with PyCharm's own tool

commit 3bd79a07ba341d4c15da9840a5cb79f41e38e2e1
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Jan 29 09:46:01 2020 +0100

    better clean target, chisquare creator

commit 1c0c9cedd366ca51447dc24b7f617a730b25095d
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Tue Jan 28 17:37:35 2020 +0100

    refactor(style): fix organisation

    Changes include:
     - Moving __version__.py info to __init__.py;
     - Adding 'reinstall' target to Makefile;
     - Removing 'requirements' file and putting its info into 'setup.py';
     - Adding functionality to nonconvex

commit dde83a930258d22d1c67d709617b7044211d884a
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Tue Jan 28 11:12:27 2020 +0100

    include changes from notebook in module

commit c8e8e2c4d065205ffb586a5954ea3b628365d30a
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Mon Jan 27 18:48:53 2020 +0100

    bump

commit 72d9fd51ca9423da5da8dbebd772cf3e1ce9726a
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Wed Jan 15 07:45:10 2020 +0100

    refactor(style): absolute imports

    Changed all import statements to absolute imports as recommended by PEP8.
    Changed `black` target to `format` in Makefile for a more obvious name.

commit fee2f69a4d99f66b2310052b3dfb63ff26c42f16
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Tue Jan 14 17:05:26 2020 +0100

    refactor(style): trying to make pyling a bit happier

    Added module docstrings to packages' __init__.py files
    Fixed variable names and import statements
    Added pylint to conda environment, and `lint` command to Makefile
    Adapted README to changes

commit 44a32c3c1f36cc89878d2cdf964bbddecc0e48cf
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Mon Jan 13 09:09:03 2020 +0100

    feat(commandline): add pid check

    New function to check if a PID exists
    Now used in the `terminate` function

commit aa4ddf5eb0c94fa1bf7bcf78b4751b56f4102119
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Sun Jan 12 20:39:54 2020 +0100

    docs(readme): tune Docker run commands

    Adds the `--rm` flag for running containers, so that someone using this won't overpopulate their installation with numerous containers.
    People used to Docker can simply ignore this flag if they know what they're doing.

commit 5c30b0ab76aca56d76c7302e05e9286de37d6f5a
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Sun Jan 12 20:36:21 2020 +0100

    fix(env files): Add numba and remove unecessary libs

    Additions:
    * Some packages have been added to include Numba into the environment
    Removals:
    * appnope and libgfortran have been removed, as they are only necessary on macOS and anyone trying to `conda install` from these files would encounter an error. On macOS, since these are a dependency conda will take care of including them when resolving the environment.

commit bca73cfd05a1fe79ae424565c6e0c0d14d145efd
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Sat Jan 11 22:25:40 2020 +0100

    Update env with sympy

commit 9b7f227852150b88e091d227a1bf33ceedf2e0b1
Merge: 9532991 0c4613d
Author: Felix Soubelet <felix.soubelet@protonmail.com>
Date:   Fri Jan 10 14:55:08 2020 +0100

    Merge branch 'release/0.1.1' into develop

+++ pyhdtoolkit/utils/cmdline.py
@@ -5,16 +5,46 @@ Created on 2019.11.06
 Utility script to help run commands and access the commandline.
 """
 
+import errno
 import os
 import signal
+import subprocess
 import sys
-from subprocess import PIPE, STDOUT, Popen
+
+from fsbox import logging_tools
+from fsbox.contexts import timeit
+
+LOGGER = logging_tools.get_logger(__name__)
 
 
 class CommandLine:
     """
     A high-level object to encapsulate the different methods for interacting with the commandline.
     """
+
+    @staticmethod
+    def check_pid_exists(pid: int) -> bool:
+        """
+        Check whether the given PID exists in the current process table.
+        :return: a boolean stating the result.
+        """
+        if pid == 0:
+            # According to "man 2 kill",  PID 0 has a special meaning: it refers to <<every process in the process
+            # group of the calling process>>, so it is best to not go any further.
+            LOGGER.info(f"PID 0 refers to 'every process in the group of calling processes', and should be untouched.")
+            return True
+        try:
+            # sending SIG 0 only checks if process has terminated, we're not actually terminating it
+            os.kill(pid, 0)
+        except OSError as error:
+            if error.errno == errno.ESRCH:  # ERROR "No such process"
+                return False
+            if error.errno == errno.EPERM:  # ERROR "Operation not permitted" -> there's a process to deny access to.
+                return True
+            # According to "man 2 kill" possible error values are (EINVAL, EPERM, ESRCH), therefore we should
+            # never get here. If so let's be explicit in considering this an error.
+            raise error
+        return True
 
     @staticmethod
     def run(command, shell: bool = True, env=None, timeout: float = None) -> tuple:
@@ -35,23 +65,29 @@ class CommandLine:
         Usage:
             run('echo hello') -> (0, b'hello\r\n')
         """
-        p = Popen(command, shell=shell, stdout=PIPE, stderr=STDOUT, env=env)
-        stdout, _ = p.communicate(timeout=timeout)
-        return p.poll(), stdout
+        with timeit(lambda spanned: LOGGER.info(f"Ran command '{command}' in a subprocess, in: {spanned}s")):
+            process = subprocess.Popen(command, shell=shell, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, env=env)
+            stdout, _ = process.communicate(timeout=timeout)
+        return process.poll(), stdout
 
     @staticmethod
     def terminate(pid: int) -> None:
         """
         Terminate process by given pid.
         On Other platforms, using os.kill with signal.SIGTERM to kill.
+        :param pid: the process ID to kill
         """
-        os.kill(pid, signal.SIGTERM)
+        if CommandLine.check_pid_exists(pid):
+            os.kill(pid, signal.SIGTERM)
+            LOGGER.info(f"Process {pid} has successfully been terminated.")
+        else:
+            LOGGER.error(f"Process with ID {pid} could not be terminated.")
 
     @staticmethod
     def get_cmdline_argv() -> list:
         """
         Get command line argv of self python process.
-        :return: a list wi
+        :return: a list with those arguments.
         """
         return sys.argv
